feat(ci): cross-OS gemini sandbox scan matrix + windows port
Add scan-gemini-matrix workflow that fans the baseline scan out across
linux × {none, docker, podman}, macos × {none, sandbox-exec}, and
windows × {none}. Each cell runs the probe directly and (on
workflow_dispatch / schedule) via gemini-cli, then uploads its report.

Probe port: split filesystem.go into _unix.go / _windows.go so
GOOS=windows compiles (unix.Access/R_OK/W_OK are POSIX-only).

Workflow design:
- Build matrix produces linux-amd64, darwin-arm64, windows-amd64.
- Scan matrix consumes them; matrix backends are an axis, not a
  separate workflow per backend.
- Token spend gated to workflow_dispatch / weekly cron; pushes only
  run the direct probes.
- gemini-cli pinned to gemini-2.5-flash-lite, retries 5xx but not 429.
- concurrency: cancel-in-progress per ref+event.
- GEMINI_API_KEY scoped to a `gemini` deployment environment with
  branch policy restricted to main and matrix/**.

Workarounds for upstream gemini-cli 0.42.0:
- SANDBOX_FLAGS='--entrypoint ""' fixes google-gemini/gemini-cli#26964
  (sandbox image's default ENTRYPOINT ["gemini"] eats the bash -c
  re-exec, producing "Unknown argument: c").
- --skip-trust on argv so the inner sandboxed gemini honours it
  (GEMINI_CLI_TRUST_WORKSPACE env var doesn't cross the re-exec).

windows-docker dropped: gemini's --sandbox image is Linux-only and
Docker Desktop on the windows-latest runner can't pull it; no native
Windows sandbox exists in gemini-cli today. That's the finding, not
a cell worth re-running each week.
